backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
langchain
langchain-core
langchain-google-genai
langgraph
pypdf
pdfplumber
pymupdf
gTTS
edge-tts
python-multipart
python-dotenv
pydantic
pydantic-settings
alembic
pytest
httpx
